moves cd adapters to deployment package
