+++ internal/deployment/evaluationfinished_adapter.go
@@ -1,4 +1,4 @@
-package adapter
+package deployment
 
 import (
 	"fmt"
